Move notif/schema tree components to their own folders
