+++ src/components/schema_trees/SchemaTree.js
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import {Tree} from "antd";
 import "antd/es/tree/style/css";
 
-import {generateSchemaTreeData} from "../schema";
+import {generateSchemaTreeData} from "../../schema";
 
 class SchemaTree extends Component {
     render() {
